backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
gunicorn==23.0.0
anthropic==0.40.0
chromadb==0.5.23
sentence-transformers>=2.2.0
PyPDF2==3.0.1
python-dotenv==1.0.1
pydantic==2.10.3
supabase>=2.0.0
PyJWT[crypto]>=2.8.0
httpx>=0.27.0
python-docx==1.1.2
python-multipart==0.0.9
sentry-sdk[fastapi]>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
